+++ packages/api-elements/test/value-of-test.js
@@ -968,6 +968,34 @@ describe('valueOf ObjectElement', () => {
     expect(value).to.deep.equal({ gaga: 'bing' });
   });
 
+  it('prefers samples over undefined property values', () => {
+    const element = new ObjectElement({ key1: new StringElement(), key2: new StringElement() });
+    element.attributes.set('default', new ObjectElement({ key1: 'defaultValue', key2: 'otherDefaultValue' }));
+    element.attributes.set('samples', new ArrayElement([
+      new ObjectElement({ key1: 'sampleValue', key2: 'otherSampleValue' }),
+    ]));
+    const value = element.valueOf();
+
+    expect(value).to.deep.equal({ key1: 'sampleValue', key2: 'otherSampleValue' });
+  });
+
+  it.only('prefers default over undefined property values', () => {
+    const defaults = new ObjectElement({ key1: 'defaultValue', key2: 'otherDefaultValue' });
+
+    const element1 = new ObjectElement({ key1: new StringElement(), key2: new StringElement() });
+    element1.attributes.set('default', defaults);
+    const value1 = element1.valueOf();
+
+    const element2 = new ObjectElement([
+      new MemberElement(), new MemberElement(),
+    ]);
+    element2.attributes.set('default', defaults);
+    const value2 = element2.valueOf();
+
+    expect(value1).to.deep.equal({ key1: 'defaultValue', key2: 'otherDefaultValue' });
+    expect(value2).to.deep.equal({ key1: 'defaultValue', key2: 'otherDefaultValue' });
+  });
+
   it('generates {} if no content, default, samples and not nullable', () => {
     const element = new ObjectElement();
     const value = element.valueOf();